+++ src/Domain/Model/Order/Payment/PaymentState.php
@@ -25,6 +25,11 @@ enum PaymentState: string implements State
         return $this->value;
     }
 
+    public function equals($other): bool
+    {
+        return (get_class($this) === get_class($other) && $this->getValueAsString() === $other->getValueAsString());
+    }
+
     public static function getDefaultTransitions(): array
     {
         return [