+++ src/Domain/Model/Order/Shipping/ShippingState.php
@@ -24,6 +24,11 @@ enum ShippingState: string implements State
         return $this->value;
     }
 
+    public function equals($other): bool
+    {
+        return (get_class($this) === get_class($other) && $this->getValueAsString() === $other->getValueAsString());
+    }
+
     public static function getDefaultTransitions(): array
     {
         return [